fix: address review findings from PR automated review
Auto-generated commit addressing issues identified in PR review.
See PR comments for detailed list of fixes applied.

Changes made via automated workflow (rite --fix-review mode).
